Refactor NAT parameters
- `nat_per_az` and `private_nat_gateway` become ternary `nat`
  parameter; compatibilty matrix looks like:
```
nat		nat_per_az	nat_private_gateway
'disabled'	False		False
'enabled'	False		True
'per-az'	True		True
```

Extending validation for `OrbitRegion` objects: enforce `nat` values as
described, validate `bastion_instance_type` is a real instance type,
etc.
