+++ sample/orbit/develop.json
@@ -5,7 +5,7 @@
   ],
   "us-east-1": {
     "bastion_instance_count": 0,
-    "private_nat_gateway": "disabled"
+    "nat": "disabled"
   },
   "us-west-2": {
     "provider": "gdh",
