+++ src/spacel/aws/__init__.py
@@ -1,3 +1,2 @@
 from .ami import AmiFinder
 from .clients import ClientCache
-from .instance_types import INSTANCE_VOLUMES
