+++ src/spacel/model/aws.py
@@ -9,6 +9,61 @@ VALID_REGIONS = (
     'ap-northeast-1'
 )
 
-INSTANCE_TYPES = (
-    't2.nano'
-)
+INSTANCE_VOLUMES = {
+    'c1.medium': 1,
+    'c1.xlarge': 4,
+    'c3.2xlarge': 2,
+    'c3.4xlarge': 2,
+    'c3.8xlarge': 2,
+    'c3.large': 2,
+    'c3.xlarge': 2,
+    'c4.2xlarge': 0,
+    'c4.4xlarge': 0,
+    'c4.8xlarge': 0,
+    'c4.large': 0,
+    'c4.xlarge': 0,
+    'cc2.8xlarge': 4,
+    'cg1.4xlarge': 2,
+    'cr1.8xlarge': 2,
+    'd2.2xlarge': 6,
+    'd2.4xlarge': 12,
+    'd2.8xlarge': 24,
+    'd2.xlarge': 3,
+    'g2.2xlarge': 1,
+    'g2.8xlarge': 2,
+    'hi1.4xlarge': 2,
+    'hs1.8xlarge': 24,
+    'i2.2xlarge': 2,
+    'i2.4xlarge': 4,
+    'i2.8xlarge': 8,
+    'i2.xlarge': 1,
+    'm1.large': 2,
+    'm1.medium': 1,
+    'm1.small': 1,
+    'm1.xlarge': 4,
+    'm2.2xlarge': 1,
+    'm2.4xlarge': 2,
+    'm2.xlarge': 1,
+    'm3.2xlarge': 2,
+    'm3.large': 1,
+    'm3.medium': 1,
+    'm3.xlarge': 2,
+    'm4.10xlarge': 0,
+    'm4.2xlarge': 0,
+    'm4.4xlarge': 0,
+    'm4.large': 0,
+    'm4.xlarge': 0,
+    'r3.2xlarge': 1,
+    'r3.4xlarge': 1,
+    'r3.8xlarge': 2,
+    'r3.large': 1,
+    'r3.xlarge': 1,
+    't1.micro': 0,
+    't2.large': 0,
+    't2.medium': 0,
+    't2.micro': 0,
+    't2.nano': 0,
+    't2.small': 0
+}
+
+INSTANCE_TYPES = set(INSTANCE_VOLUMES.keys())
